Handle gcc quirk with parameter checking
At least some versions of GCC appear to take any -Wno-foo option
without error even if that option is not supported.  But they will
error out if -Wno-error=foo is used.  This sounds like a GCC bug, but
we'll work around it and test for the feature both ways.

Thanks-to: Matteo Brancaleoni <mbrancaleoni@voismart.it>

FS-6850 #resolve